refactor: Extract `deployment-config`
