+++ Cargo.toml
@@ -12,6 +12,7 @@ version.workspace = true
 
 [dependencies]
 concepts.workspace = true
+deployment-config.workspace = true
 db-postgres.workspace = true
 db-sqlite.workspace = true
 executor.workspace = true
@@ -146,6 +147,7 @@ members = [
     "crates/bench",
     "crates/concepts",
     "crates/db-common",
+    "crates/deployment-config",
     "crates/db-postgres",
     "crates/db-sqlite",
     "crates/executor",
@@ -206,6 +208,7 @@ concepts = { package = "obeli-sk-concepts", path = "crates/concepts", version = 
 db-common = { package = "obeli-sk-db-common", path = "crates/db-common", version = "0.38.3" }
 db-postgres = { package = "obeli-sk-db-postgres", path = "crates/db-postgres", version = "0.38.3" }
 db-sqlite = { package = "obeli-sk-db-sqlite", path = "crates/db-sqlite", version = "0.38.3" }
+deployment-config = { package = "obeli-sk-deployment-config", path = "crates/deployment-config", version = "0.38.3" }
 db-tests = { package = "obeli-db-tests", path = "crates/testing/db-tests" }
 executor = { package = "obeli-sk-executor", path = "crates/executor", version = "0.38.3" }
 grpc = { package = "obeli-sk-grpc", path = "crates/grpc", version = "0.38.3" }
