+++ crates/concepts/src/rusqlite_ext.rs
@@ -1,6 +1,5 @@
 use crate::{
-    ComponentType, ExecutionId, FunctionFqn, JoinSetId,
-    component_id::{ComponentDigest, Digest},
+    ExecutionId, JoinSetId,
     prefixed_ulid::{DelayId, DeploymentId, ExecutionIdDerived, ExecutorId, RunId},
     storage::{
         DbErrorGeneric, DbErrorRead, DbErrorReadWithTimeout, DbErrorWrite, DbErrorWriteNonRetriable,
@@ -10,7 +9,7 @@ use rusqlite::{
     ErrorCode, ToSql,
     types::{FromSql, FromSqlError, FromSqlResult, ToSqlOutput, ValueRef},
 };
-use std::{panic::Location, str::FromStr, sync::Arc};
+use std::{panic::Location, sync::Arc};
 use tracing::error;
 use tracing_error::SpanTrace;
 
@@ -113,29 +112,6 @@ impl FromSql for ExecutorId {
     }
 }
 
-impl ToSql for ComponentDigest {
-    fn to_sql(&self) -> rusqlite::Result<rusqlite::types::ToSqlOutput<'_>> {
-        Ok(ToSqlOutput::from(self.as_slice()))
-    }
-}
-impl FromSql for ComponentDigest {
-    fn column_result(value: ValueRef<'_>) -> FromSqlResult<Self> {
-        match value {
-            ValueRef::Blob(b) => {
-                let digest = Digest::try_from(b).map_err(|err| {
-                    error!(
-                        "Cannot convert to {} - {err:?}",
-                        std::any::type_name::<Self>()
-                    );
-                    FromSqlError::Other(Box::from(err.to_string()))
-                })?;
-                Ok(ComponentDigest(digest))
-            }
-            _ => Err(FromSqlError::InvalidType),
-        }
-    }
-}
-
 impl From<rusqlite::Error> for DbErrorRead {
     #[track_caller]
     fn from(err: rusqlite::Error) -> Self {
@@ -177,52 +153,12 @@ impl From<rusqlite::Error> for DbErrorWrite {
     }
 }
 
-impl FromSql for FunctionFqn {
-    #[track_caller]
-    fn column_result(value: ValueRef<'_>) -> FromSqlResult<Self> {
-        match value {
-            ValueRef::Text(ffqn) => {
-                let ffqn = String::from_utf8_lossy(ffqn);
-                let ffqn = FunctionFqn::from_str(&ffqn).map_err(|err| {
-                    let err = DbErrorGeneric::Uncategorized {
-                        reason: "Cannot deserialize ffqn".into(),
-                        context: SpanTrace::capture(),
-                        source: Some(Arc::new(err)),
-                        loc: Location::caller(),
-                    };
-                    FromSqlError::Other(Box::from(err))
-                })?;
-                Ok(ffqn)
-            }
-            _ => Err(FromSqlError::InvalidType),
-        }
-    }
-}
-
 impl ToSql for DeploymentId {
     fn to_sql(&self) -> rusqlite::Result<rusqlite::types::ToSqlOutput<'_>> {
         Ok(ToSqlOutput::from(self.to_string()))
     }
 }
 impl FromSql for DeploymentId {
-    fn column_result(value: ValueRef<'_>) -> FromSqlResult<Self> {
-        let str = value.as_str()?;
-        str.parse::<Self>().map_err(|err| {
-            error!(
-                "Cannot convert to {} value:`{str}` - {err:?}",
-                std::any::type_name::<Self>()
-            );
-            FromSqlError::InvalidType
-        })
-    }
-}
-
-impl ToSql for ComponentType {
-    fn to_sql(&self) -> rusqlite::Result<rusqlite::types::ToSqlOutput<'_>> {
-        Ok(ToSqlOutput::from(self.to_string()))
-    }
-}
-impl FromSql for ComponentType {
     fn column_result(value: ValueRef<'_>) -> FromSqlResult<Self> {
         let str = value.as_str()?;
         str.parse::<Self>().map_err(|err| {
