+++ crates/deployment-config/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "obeli-sk-deployment-config"
+description = "Shared deployment configuration schema and core naming types of obelisk. Compiles to wasm32-unknown-unknown for the webui."
+
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+arbitrary = { workspace = true, optional = true }
+const_format.workspace = true
+derive_more.workspace = true
+hashbrown.workspace = true
+rusqlite = { workspace = true, optional = true }
+schemars.workspace = true
+serde.workspace = true
+serde_with.workspace = true
+strum.workspace = true
+thiserror.workspace = true
+tokio-postgres = { workspace = true, optional = true }
+tracing.workspace = true
+
+[features]
+test = ["dep:arbitrary"]
+rusqlite = ["dep:rusqlite"]
+postgres = ["dep:tokio-postgres"]
+
+[lints]
+workspace = true
