+++ src/command/server.rs
@@ -5,32 +5,44 @@ use crate::config::config_holder::ConfigHolder;
 use crate::config::config_holder::PathPrefixes;
 use crate::config::config_holder::load_deployment_canonical;
 use crate::config::env_var::EnvVarConfig;
+use crate::config::toml::ActivityExecComponentConfigCanonicalExt as _;
 use crate::config::toml::ActivityExecConfigVerified;
+use crate::config::toml::ActivityExternalComponentConfigCanonicalExt as _;
 use crate::config::toml::ActivityExternalConfigVerified;
+use crate::config::toml::ActivityJsComponentConfigCanonicalExt as _;
 use crate::config::toml::ActivityJsConfigVerified;
+use crate::config::toml::ActivityStubComponentConfigCanonicalExt as _;
 use crate::config::toml::ActivityStubConfigVerified;
 use crate::config::toml::ActivityStubExtConfigVerified;
 use crate::config::toml::ActivityStubExtInlineConfigVerified;
+use crate::config::toml::ActivityWasmComponentConfigTomlExt as _;
 use crate::config::toml::ActivityWasmConfigVerified;
 use crate::config::toml::CancelWatcherTomlConfig;
 use crate::config::toml::ComponentCommon;
+use crate::config::toml::ComponentLocationFetchExt as _;
 use crate::config::toml::ComponentStdOutputToml;
 use crate::config::toml::ConfigName;
 use crate::config::toml::DatabaseConfigToml;
 use crate::config::toml::DeploymentCanonical;
+use crate::config::toml::InflightSemaphoreExt as _;
 use crate::config::toml::LogLevelToml;
 use crate::config::toml::SQLITE_FILE_NAME;
 use crate::config::toml::ServerConfigToml;
 use crate::config::toml::TimersWatcherTomlConfig;
 use crate::config::toml::WasmtimeAllocatorConfig;
 use crate::config::toml::WorkflowConfigVerified;
+use crate::config::toml::WorkflowJsComponentConfigCanonicalExt as _;
 use crate::config::toml::WorkflowJsConfigVerified;
+use crate::config::toml::WorkflowWasmComponentConfigCanonicalExt as _;
+use crate::config::toml::cron::CronComponentConfigTomlExt as _;
 use crate::config::toml::cron::CronConfigVerified;
 use crate::config::toml::webhook;
 use crate::config::toml::webhook::HttpServer;
+use crate::config::toml::webhook::WebhookJsComponentConfigCanonicalExt as _;
 use crate::config::toml::webhook::WebhookJsConfigVerified;
 use crate::config::toml::webhook::WebhookRoute;
 use crate::config::toml::webhook::WebhookRouteVerified;
+use crate::config::toml::webhook::WebhookWasmComponentConfigCanonicalExt as _;
 use crate::config::toml::webhook::WebhookWasmComponentConfigVerified;
 use crate::config::toml::{AllowedHostToml, MethodsInput, MethodsInputStar};
 use crate::config::wasm_cache_metadata_dir;
