+++ src/config/env_var.rs
@@ -1,19 +1,6 @@
-use schemars::JsonSchema;
 use secrecy::SecretString;
-use serde::{Deserialize, Serialize};
 
-#[derive(Clone, derive_more::Debug, Hash, JsonSchema, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum EnvVarConfig {
-    /// Forward from host: `"KEY"`
-    Key(String),
-    /// Set to value: `{key = "KEY", value = "foo"}` (supports `${VAR}` interpolation)
-    KeyValue {
-        key: String,
-        #[debug(skip)]
-        value: String,
-    },
-}
+pub use deployment_config::env_var::EnvVarConfig;
 
 #[derive(Debug, thiserror::Error)]
 #[error("environment variable not set: `{0}`")]
